requirements.txt
asgiref==3.11.1
beautifulsoup4==4.12.3
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.2.11
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.11
lxml==5.3.0
pillow==11.0.0
PyJWT==2.11.0
python-decouple==3.8
curl-cffi>=0.7
playwright>=1.40
requests==2.32.3
soupsieve==2.8.3
sqlparse==0.5.5
urllib3==2.6.3
